# backend/golang/internal/api/docs/swagger.yaml
definitions:
  dtos.Appointment:
    properties:
      customerId:
        type: string
      customerName:
        type: string
      date:
        type: string
      duration:
        type: integer
      id:
        type: string
      price:
        type: integer
      serviceName:
        type: string
      serviceNameId:
        type: string
      specialistId:
        type: string
      specialistName:
        type: string
      status:
        description: 'TODO: remove'
        type: integer
      time:
        type: string
    type: object
  dtos.AppointmentCalendar:
    properties:
      date:
        type: string
      id:
        type: string
      specialistName:
        type: string
      status:
        type: integer
      time:
        type: string
    type: object
  dtos.AppointmentCalendarCount:
    properties:
      canceledCount:
        type: integer
      month:
        type: integer
      pendingCount:
        type: integer
      realizedCount:
        type: integer
    type: object
  dtos.CreateAppointmentBody:
    properties:
      customerId:
        type: string
      date:
        type: string
      serviceId:
        type: string
      time:
        type: string
    type: object
...
  dtos.UpdateAppointmentBody:
    properties:
      date:
        type: string
      status:
        type: integer
      time:
        type: string
    type: object
info:
  contact: {}
paths:
  /appointments:
    get:
      description: Get a page of appointments with filters sorted by date and time
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Start date
        in: query
        name: startDate
        type: string
      - description: End date
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.Appointment'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List appointments
      tags:
      - Appointments
    post:
      consumes:
      - application/json
      description: Create a new appointment for a customer and service
      parameters:
      - description: Appointments info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateAppointmentBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create appointment
      tags:
      - Appointments
  /appointments/{appointment_id}:
    delete:
      description: Delete an appointment by id
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete appointment
      tags:
      - Appointments
    get:
      description: Get an appointment by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Appointment'
      security:
      - ApiKeyAuth: []
      summary: Get appointment
      tags:
      - Appointments
    put:
      consumes:
      - application/json
      description: Create a new appointment for a customer and service
      parameters:
      - description: Appointments info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateAppointmentBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create appointment
      tags:
      - Appointments
  /appointments/calendar:
    get:
      description: Get appointments for calendar in a interval of date
      parameters:
      - description: Year
        in: query
        name: year
        type: integer
      - description: Start date
        in: query
        name: startDate
        type: string
      - description: End date
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.AppointmentCalendar'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List appointments for calendar
      tags:
      - Appointments
  /appointments/calendar/count:
    get:
      description: Get appointments count for calendar
      parameters:
      - description: Year
        in: query
        name: year
        type: integer
      - description: Start date
        in: query
        name: startDate
        type: string
      - description: End date
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.AppointmentCalendarCount'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Count appointments for calendar
      tags:
      - Appointments
  /appointments/count:
    get:
      description: Count appointments with filters
      parameters:
      - description: Start date
        in: query
        name: startDate
        type: string
      - description: End date
        in: query
        name: endDate
        type: string
      - description: Service Name ID
        in: query
        name: serviceNameId
        type: string
      - description: Specialist ID
        in: query
        name: specialistId
        type: string
      - description: Customer ID
        in: query
        name: customerId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Count appointments
      tags:
      - Appointments
...
swagger: "2.0"